Aral - changes in create project presenter, create project view completed
