--- src/main/java/view/project/ProjectView.java
@@ -1,4 +1,0 @@
-package view.project;
-
-public class ProjectView {
-}
